:recycle: migrate tests about TupleView size
