+++ Tests/SwiftTUITests/Primitive/TupleViewTests.swift
@@ -1,0 +1,43 @@
+//
+//  TupleViewTests.swift
+//  SwiftTUITests
+//
+//  Created by Yudai.Hirose on 2020/03/31.
+//
+
+import XCTest
+@testable import SwiftTUI
+
+class TupleViewTests: XCTestCase {
+    override func setUp() {
+        super.setUp()
+        
+        mainScreen = DummyScreen.init()
+    }
+    
+    func testSize() {
+        func prepare<T: View>(view: T, viewListOption: ViewVisitorListOption = .vertical) -> ViewGraph {
+            let graphVisitor = ViewGraphSetVisitor()
+            let graph = graphVisitor.visit(view)
+            graph.listType = viewListOption
+            return graph
+        }
+        
+        XCTContext.runActivity(named: "when TupleView<Text, Text, Text>") { (_) in
+            let view = TupleView((
+                Text("1"),
+                Text("23"),
+                Text("456")
+            ))
+            
+            let graph = prepare(view: view)
+            let visitor = ViewSetRectVisitor()
+            graph.accept(visitor: visitor)
+            
+            let elementCount = 3
+            let spacing = (elementCount - 1) * ViewVisitorListOption.vertical.defaultSpace
+            
+            XCTAssertEqual(graph.rect.size, Size(width: "456".width, height: elementCount + spacing))
+        }
+    }
+}
